Move BuildGcsfuse into util package
